feature(app): Code Clean
issue #40
