+++ app/src/main/java/com/despaircorp/realestatemanagerkotlin/main/legacy/Utils.java
@@ -1,0 +1,49 @@
+package com.despaircorp.realestatemanagerkotlin.main.legacy;
+
+import android.content.Context;
+import android.net.wifi.WifiManager;
+
+import java.text.DateFormat;
+import java.text.SimpleDateFormat;
+import java.util.Date;
+
+/**
+ * Created by Philippe on 21/02/2018.
+ */
+
+public class Utils {
+
+    /**
+     * Conversion d'un prix d'un bien immobilier (Dollars vers Euros)
+     * NOTE : NE PAS SUPPRIMER, A MONTRER DURANT LA SOUTENANCE
+     *
+     * @param dollars
+     * @return
+     */
+    public static int convertDollarToEuro(int dollars) {
+        return (int) Math.round(dollars * 0.812);
+    }
+
+    /**
+     * Conversion de la date d'aujourd'hui en un format plus approprié
+     * NOTE : NE PAS SUPPRIMER, A MONTRER DURANT LA SOUTENANCE
+     *
+     * @return
+     */
+    public static String getTodayDate() {
+        DateFormat dateFormat = new SimpleDateFormat("dd/MM/yyyy");
+        return dateFormat.format(new Date());
+    }
+
+    /**
+     * Vérification de la connexion réseau
+     * NOTE : NE PAS SUPPRIMER, A MONTRER DURANT LA SOUTENANCE
+     *
+     * @param context
+     * @return
+     */
+    public static Boolean isInternetAvailable(Context context) {
+        WifiManager wifi = (WifiManager) context.getSystemService(Context.WIFI_SERVICE);
+        return wifi.isWifiEnabled();
+    }
+}
