+++ data/build.gradle.kts
@@ -62,7 +62,7 @@ dependencies {
     implementation("com.google.android.gms:play-services-maps:18.2.0")
     implementation("com.google.android.gms:play-services-location:21.2.0")
     testImplementation(project(":stubs"))
-    
+    implementation("org.chromium.net:cronet-embedded:119.6045.31")
     coreLibraryDesugaring("com.android.tools:desugar_jdk_libs:2.0.4")
     implementation("com.google.code.gson:gson:2.10.1")
     implementation("com.google.dagger:hilt-android:2.48.1")
