+++ domain/src/test/java/com/despaircorp/domain/picture/InsertPictureForEstateCreationUseCaseUnitTest.kt
@@ -28,7 +28,7 @@ class InsertPictureForEstateCreationUseCaseUnitTest {
     
     @Test
     fun `nominal case - insert picture`() = testCoroutineRule.runTest {
-        useCase.invoke(EntityProvidinator.provideEstatePicturesEntities(), DEFAULT_ID, isEditMode)
+        useCase.invoke(EntityProvidinator.provideEstatePicturesEntities(), DEFAULT_ID, false)
         
         coVerify {
             pictureDomainRepository.insertPictures(EntityProvidinator.provideEstatePictureEntity())
