+++ shared/build.gradle.kts
@@ -33,7 +33,7 @@ android {
 }
 
 dependencies {
-    
+    implementation("org.chromium.net:cronet-embedded:119.6045.31")
     implementation("androidx.core:core-ktx:1.12.0")
     implementation("androidx.appcompat:appcompat:1.6.1")
     implementation("com.google.android.material:material:1.11.0")
