+++ stubs/build.gradle.kts
@@ -45,6 +45,7 @@ dependencies {
     coreLibraryDesugaring("com.android.tools:desugar_jdk_libs:2.0.4")
     implementation("androidx.appcompat:appcompat:1.6.1")
     implementation("com.google.android.material:material:1.11.0")
+    implementation("org.chromium.net:cronet-embedded:119.6045.31")
     implementation("androidx.constraintlayout:constraintlayout:2.1.4")
     testImplementation("androidx.arch.core:core-testing:2.1.0") {
         // Removes the Mockito dependency bundled with arch core (wtf android ??)
