+++ ui/src/main/java/com/despaircorp/ui/main/master_fragment/MasterViewModel.kt
@@ -1,7 +1,6 @@
 package com.despaircorp.ui.main.master_fragment
 
 import android.app.Application
-import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.liveData
 import androidx.sqlite.db.SimpleSQLiteQuery
@@ -9,13 +8,14 @@ import com.despaircorp.domain.connectivity.IsUserConnectedToInternetUseCase
 import com.despaircorp.domain.currency.GetActualCurrencyUseCase
 import com.despaircorp.domain.currency.model.CurrencyEnum
 import com.despaircorp.domain.estate.GetEstateWithPictureEntityAsFlowUseCase
-import com.despaircorp.domain.estate.model.EstateStatus
 import com.despaircorp.domain.estate.model.EstateTypeEntity
 import com.despaircorp.domain.estate.model.PointOfInterestEntity
+import com.despaircorp.domain.filter.EstateSqlQueryBuilderUseCase
+import com.despaircorp.domain.filter.FilterTypeEnum
+import com.despaircorp.domain.filter.OnFilterChangedUseCase
 import com.despaircorp.domain.geocoder.GetAddressFromLatLngUseCase
 import com.despaircorp.ui.R
 import com.despaircorp.ui.main.master_fragment.estate.EstateViewStateItems
-import com.despaircorp.ui.main.master_fragment.filter.FilterTypeEnum
 import com.despaircorp.ui.utils.getEuroFromDollar
 import com.google.android.gms.maps.model.LatLng
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -33,10 +33,12 @@ class MasterViewModel @Inject constructor(
     private val getAddressFromLatLngUseCase: GetAddressFromLatLngUseCase,
     private val application: Application,
     private val getActualCurrencyUseCase: GetActualCurrencyUseCase,
-    private val isUserConnectedToInternetUseCase: IsUserConnectedToInternetUseCase
+    private val isUserConnectedToInternetUseCase: IsUserConnectedToInternetUseCase,
+    private val onFilterChangedUseCase: OnFilterChangedUseCase,
+    private val estateSqlQueryBuilderUseCase: EstateSqlQueryBuilderUseCase
 ) : ViewModel() {
     
-    private val selectedItem = MutableStateFlow(1)
+    private val selectedItemMutableStateFlow = MutableStateFlow(1)
     
     private val pointOfInterestEntitiesMutableStateFlow =
         MutableStateFlow(emptyList<PointOfInterestEntity>())
@@ -44,7 +46,8 @@ class MasterViewModel @Inject constructor(
         MutableStateFlow(emptyList<EstateTypeEntity>())
     
     private val rawQueryMutableStateFlow =
-        MutableStateFlow("SELECT * FROM estate_table")
+        MutableStateFlow(BASE_SQL_QUERY)
+    
     private val toBuildQueryMutableStateFlow = MutableStateFlow(
         (mutableMapOf<String, String>(
             "surfaceMin" to "",
@@ -62,9 +65,9 @@ class MasterViewModel @Inject constructor(
     val viewState = liveData<MasterViewState> {
         
         rawQueryMutableStateFlow.collectLatest { rawQuery ->
-            Log.i("Monokouma", rawQuery)
+            
             combine(
-                selectedItem,
+                selectedItemMutableStateFlow,
                 getActualCurrencyUseCase.invoke(),
                 isUserConnectedToInternetUseCase.invoke(),
                 getEstateWithPictureEntityAsFlowUseCase.invoke(SimpleSQLiteQuery(rawQuery))
@@ -128,105 +131,19 @@ class MasterViewModel @Inject constructor(
     }
     
     fun onEstateClicked(estateId: Int) {
-        selectedItem.value = estateId
+        selectedItemMutableStateFlow.value = estateId
     }
     
     fun onFilterChanged(filterArg: String, filterBy: FilterTypeEnum) {
         //React with enum etc
-        Log.i("Monokouma", filterBy.toString())
-        when (filterBy) {
-            FilterTypeEnum.SURFACE_MIN -> {
-                
-                val currentMap = toBuildQueryMutableStateFlow.value.toMutableMap()
-                if (filterArg.isEmpty()) {
-                    currentMap["surfaceMin"] = ""
-                    toBuildQueryMutableStateFlow.value = currentMap
-                } else {
-                    currentMap["surfaceMin"] = filterArg
-                    toBuildQueryMutableStateFlow.value = currentMap
-                }
-            }
-            
-            FilterTypeEnum.SURFACE_MAX -> {
-                
-                val currentMap = toBuildQueryMutableStateFlow.value.toMutableMap()
-                if (filterArg.isEmpty()) {
-                    currentMap["surfaceMax"] = ""
-                    toBuildQueryMutableStateFlow.value = currentMap
-                } else {
-                    currentMap["surfaceMax"] = filterArg
-                    toBuildQueryMutableStateFlow.value = currentMap
-                }
-            }
-            
-            FilterTypeEnum.PRICE_MIN -> {
-                val currentMap = toBuildQueryMutableStateFlow.value.toMutableMap()
-                if (filterArg.isEmpty()) {
-                    currentMap["priceMin"] = ""
-                    toBuildQueryMutableStateFlow.value = currentMap
-                } else {
-                    currentMap["priceMin"] = filterArg
-                    toBuildQueryMutableStateFlow.value = currentMap
-                }
-            }
-            
-            FilterTypeEnum.PRICE_MAX -> {
-                val currentMap = toBuildQueryMutableStateFlow.value.toMutableMap()
-                if (filterArg.isEmpty()) {
-                    currentMap["priceMax"] = ""
-                    toBuildQueryMutableStateFlow.value = currentMap
-                } else {
-                    currentMap["priceMax"] = filterArg
-                    toBuildQueryMutableStateFlow.value = currentMap
-                }
-            }
-            
-            FilterTypeEnum.ROOM_NUMBER -> {
-                val currentMap = toBuildQueryMutableStateFlow.value.toMutableMap()
-                if (filterArg == "roomNumber=''") {
-                    currentMap["roomNumber"] = ""
-                    toBuildQueryMutableStateFlow.value = currentMap
-                } else {
-                    currentMap["roomNumber"] = filterArg
-                    toBuildQueryMutableStateFlow.value = currentMap
-                }
-            }
-            
-            FilterTypeEnum.FOR_SALE -> {
-                val currentMap = toBuildQueryMutableStateFlow.value.toMutableMap()
-                currentMap["sold"] = ""
-                currentMap["forSale"] = filterArg
-                toBuildQueryMutableStateFlow.value = currentMap
-            }
-            
-            FilterTypeEnum.SOLD -> {
-                val currentMap = toBuildQueryMutableStateFlow.value.toMutableMap()
-                currentMap["sold"] = filterArg
-                currentMap["forSale"] = ""
-                toBuildQueryMutableStateFlow.value = currentMap
-            }
-            
-            FilterTypeEnum.ENTRY_DATE -> {
-                val currentMap = toBuildQueryMutableStateFlow.value.toMutableMap()
-                currentMap["entryDate"] = filterArg
-                toBuildQueryMutableStateFlow.value = currentMap
-            }
-            
-            FilterTypeEnum.POINT_OF_INTEREST -> {
-                TODO()
-            }
-            
-            FilterTypeEnum.ESTATE_TYPE -> {
-                TODO()
-            }
-        }
-        
+        toBuildQueryMutableStateFlow.value =
+            onFilterChangedUseCase.invoke(filterArg, filterBy, toBuildQueryMutableStateFlow.value)
     }
     
     fun onApplyFilter() {
         val queryMap = toBuildQueryMutableStateFlow.value
         
-        val sqlQuery = buildEstateSearchQuery(
+        val sqlQuery = estateSqlQueryBuilderUseCase.invoke(
             surfaceMin = queryMap["surfaceMin"],
             surfaceMax = queryMap["surfaceMax"],
             priceMin = queryMap["priceMin"],
@@ -238,77 +155,12 @@ class MasterViewModel @Inject constructor(
             estateType = estateTypeListMutableStateFlow.value,
             pointsOfInterest = pointOfInterestEntitiesMutableStateFlow.value
         )
+        
         rawQueryMutableStateFlow.value = sqlQuery
-        Log.i("Monokouma", sqlQuery)
-    }
-    
-    private fun buildEstateSearchQuery(
-        surfaceMin: String?,
-        surfaceMax: String?,
-        priceMin: String?,
-        priceMax: String?,
-        roomNumber: String?,
-        forSale: String?,
-        sold: String?,
-        entryDate: String?,
-        estateType: List<EstateTypeEntity>?, // Assuming this is a single value, you might need to adjust if multiple types can be selected
-        pointsOfInterest: List<PointOfInterestEntity>? // List of selected points of interest like schools, parks, etc.
-    ): String {
-        
-        val queryBuilder = StringBuilder("SELECT * FROM estate_table WHERE 1=1")
-        
-        // Surface conditions
-        surfaceMin?.takeIf { it.isNotBlank() }?.also { queryBuilder.append(" AND surface >= $it") }
-        surfaceMax?.takeIf { it.isNotBlank() }?.also { queryBuilder.append(" AND surface <= $it") }
-        
-        // Price conditions
-        when {
-            !priceMin.isNullOrBlank() && !priceMax.isNullOrBlank() -> {
-                queryBuilder.append(" AND price BETWEEN ${priceMin.toInt()} AND ${priceMax.toInt()}")
-            }
-            
-            !priceMin.isNullOrBlank() -> {
-                queryBuilder.append(" AND price >= ${priceMin.toInt()}")
-            }
-            
-            !priceMax.isNullOrBlank() -> {
-                queryBuilder.append(" AND price <= ${priceMax.toInt()}")
-            }
-        }
-        
-        
-        // Room number condition
-        roomNumber?.takeIf { it.isNotBlank() }
-            ?.also { queryBuilder.append(" AND roomNumber = $it") }
-        
-        // Sale status conditions
-        forSale?.takeIf { it.isNotBlank() }
-            ?.also { queryBuilder.append(" AND status = '${EstateStatus.FOR_SALE.name}'") }
-        sold?.takeIf { it.isNotBlank() }
-            ?.also { queryBuilder.append(" AND status = '${EstateStatus.SOLD.name}'") }
-        
-        // Entry date condition
-        entryDate?.takeIf { it.isNotBlank() }
-            ?.also { queryBuilder.append(" AND entryDate = '$it'") }
-        
-        // Points of interest conditions
-        if (pointsOfInterest?.isNotEmpty() == true) {
-            val poiConditions =
-                pointsOfInterest.joinToString(separator = " OR ") { "pointOfInterest LIKE '%${it.pointOfInterestEnum.name}%'" }
-            queryBuilder.append(" AND ($poiConditions)")
-        }
-        Log.i("Monokouma", estateType.toString())
-        if (estateType?.isNotEmpty() == true) {
-            val estateTypeConditions =
-                estateType.joinToString(separator = " OR ") { "estateType LIKE '%${it.estateTypeEnum.name}%'" }
-            queryBuilder.append(" AND ($estateTypeConditions)")
-        }
-        
-        return queryBuilder.toString()
     }
     
     fun onResetFilter() {
-        rawQueryMutableStateFlow.value = "SELECT * FROM estate_table"
+        rawQueryMutableStateFlow.value = BASE_SQL_QUERY
         toBuildQueryMutableStateFlow.value = mutableMapOf(
             "surfaceMin" to "",
             "surfaceMax" to "",
@@ -329,5 +181,9 @@ class MasterViewModel @Inject constructor(
     fun onEstateTypeForFilteringChanged(estateTypeEntities: List<EstateTypeEntity>) {
         estateTypeListMutableStateFlow.value =
             estateTypeEntities.filter { it.isSelected }
+    }
+    
+    companion object {
+        private const val BASE_SQL_QUERY = "SELECT * FROM estate_table"
     }
 }
